pyproject.toml
[project]
name = "job-hunter"
version = "0.1.0"
description = "Agentic AI job hunter for Indian job platforms"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "openai>=1.60.0",
    "groq>=0.15.0",
    "playwright>=1.49.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "rich>=13.9.0",
    "click>=8.1.0",
    "aiosqlite>=0.21.0",
    "pypdf>=5.1.0",
    "python-docx>=1.1.0",
]

[project.scripts]
job-hunter = "job_hunter.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]
